+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog
+
+## 1.0.4
+
+- Restructured the logic of copying files/folders from the template.
+- Ensures `lib/` and subfolders are copied correctly.
+- Deletes old destination folders before copying new ones.
+- Prepares assets (`assets/images` and `assets/svg`) automatically.
+
 ## 1.0.3
 
 - Initial version.